+++ phtree/phtree_multimap.h
@@ -413,11 +413,19 @@ class PhTreeMultiMap {
      * @param new_key The new position
      * @param value The value that needs to be relocated. The relocate() method used the value's
      *              '==' operator to identify the entry that should be moved.
+     * @param count_equals This setting toggles whether a relocate() between two identical keys
+     *              should be counted as 'success' and return '1'. The function may still return '0'
+     *              in case the keys are not in the index.
+     *              Background: the intuitively correct behavior is to return '1' for identical
+     *              (exising) keys. However, avoiding this check can considerably speed up
+     *              relocate() calls, especially when using a ConverterMultiply.
+     *
      * @return '1' if a value was found and reinserted, otherwise '0'.
      */
     template <typename T2>
-    size_t relocate(const Key& old_key, const Key& new_key, T2&& value) {
-        auto pair = tree_._find_or_create_two_mm(converter_.pre(old_key), converter_.pre(new_key));
+    size_t relocate(const Key& old_key, const Key& new_key, T2&& value, bool count_equals = true) {
+        auto pair = tree_._find_or_create_two_mm(
+            converter_.pre(old_key), converter_.pre(new_key), count_equals);
         auto& iter_old = pair.first;
         auto& iter_new = pair.second;
 
@@ -469,11 +477,20 @@ class PhTreeMultiMap {
      * @param new_key The new position
      * @param predicate The predicate that is used for every value at position old_key to evaluate
      *             whether it should be relocated to new_key.
+     * @param count_equals This setting toggles whether a relocate() between two identical keys
+     *              should be counted as 'success' and return '1'. The function may still return '0'
+     *              in case the keys are not in the index.
+     *              Background: the intuitively correct behavior is to return '1' for identical
+     *              (exising) keys. However, avoiding this check can considerably speed up
+     *              relocate() calls, especially when using a ConverterMultiply.
+     *
      * @return the number of values that were relocated.
      */
     template <typename PREDICATE>
-    size_t relocate_if(const Key& old_key, const Key& new_key, PREDICATE&& predicate) {
-        auto pair = tree_._find_or_create_two_mm(converter_.pre(old_key), converter_.pre(new_key));
+    size_t relocate_if(
+        const Key& old_key, const Key& new_key, PREDICATE&& predicate, bool count_equals = true) {
+        auto pair = tree_._find_or_create_two_mm(
+            converter_.pre(old_key), converter_.pre(new_key), count_equals);
         auto& iter_old = pair.first;
         auto& iter_new = pair.second;